+++ app/models/wallet.py
@@ -7,8 +7,8 @@ class Wallet(db.Model):
 
   id = db.Column(db.Integer, primary_key=True)
   name = db.Column(db.String(40), nullable=False)
-  balance = db.Column(db.Integer, nullable=False)
-  startingBalance = db.Column(db.Integer, nullable=False)
+  balance = db.Column(db.Float, nullable=False)
+  startingBalance = db.Column(db.Float, nullable=False)
   user_id = db.Column(db.Integer, db.ForeignKey('users.id'), nullable=False)
 
   transactions = db.relationship('Transaction', lazy=True)  # DONE
